PSP seeds: 64-bit randint; document optimized mod builds on Windows
- test_syn_Fig7.py draws its synapse seeds with randint(0, 2**32 - 1) on
  its own RandomState; on Windows the default dtype is 32-bit and this
  fails. The calls now pass dtype=np.int64 (same numbers as before on
  Linux).
- NEURON 9.0.2's nrnivmodl on Windows compiles the mod files without -O;
  the PSP simulations ran about 4 times slower (about 270 s vs 72 s for the first
  connection). The Windows instructions (requirements_neuron9_windows.txt,
  README, CHANGES_NEURON9.md) now set MAKEFLAGS=EXTRA_FLAGS=-O2 first.

+++ Single Cell Modelling/PSP Simulations/test_syn_Fig7.py
@@ -424,15 +424,16 @@ for aidx, agegroup in enumerate(agegroups):
 				syndists.append(h.distance(locx))
 				
 				# Replicates LFPy's method but with control over seeds
-				rseed = local_state.randint(0,2**32 - 1)
+				# (dtype=np.int64: the default is 32-bit on Windows, too small for 2**32 - 1; same numbers as on Linux)
+				rseed = local_state.randint(0,2**32 - 1,dtype=np.int64)
 				rng = neuron.h.Random(rseed)
 				rng.MCellRan4(
 					local_state.randint(
 						0,
-						2**32 - 1),
+						2**32 - 1,dtype=np.int64),
 					local_state.randint(
 						0,
-						2**32 - 1))
+						2**32 - 1,dtype=np.int64))
 				rng.uniform(0, 1)
 				network.populations[postc].cells[0].netconsynapses[l].setRNG(rng)
 				network.populations[postc].cells[0].rng_list[l] = rng
